requirements.txt
python-dotenv==1.1.1
psycopg[binary]==3.2.10
openai==0.28.1
requests==2.31.0
